third_party/blink/web_tests/fast/dom/Element/client-rect-list-argument-expected.txt

Tests that the ClientRectList.item() argument is correctly validated.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS clientRects.length is 3
PASS clientRects.item(0) is not null
PASS clientRects.item(1) is not null
PASS clientRects.item(2) is not null
PASS clientRects.item() threw exception TypeError: Failed to execute 'item' on 'DOMRectList': 1 argument required, but only 0 present..
PASS clientRects.item(-4294967294) is clientRects.item(2)
PASS clientRects.item(3) is null
PASS clientRects.item(999) is null
PASS clientRects.item(-1) is null
PASS successfullyParsed is true

TEST COMPLETE


14. Block in inline

Lorem
ipsum
